mtd: nand: sunxi: drop subpage read support
Subpage read support has proven to be inefficient when used with UBIFS
because it prevents the usage of the NAND page cache which UBIFS
implicitely rely on to efficiently read consecutive UBIFS nodes.

Drop subpage read support until we find a better solution (page caching
at the NAND or UBI level).

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>