UBI: adapt ubi_eba_atomic_leb_change() to properly handle invalidated LEBs
LEB consolidation can lead to some LEBs being invalidated but still present
on the flash. If the newer LEB version is unmapped, then the previous
invalid LEB becomes valid again at next boot.

This can lead to some problems when the UBI user expects the unmapped LEB
to either point to the lastest version (the one scheduled for erasure) or
an empty LEB (all data set to 0xff).

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>